2002/12/17 ver 1.1.4 released
	From Bart
	- Fix X font handling and misc other things in the dosemu script 
	  and make install.
	- add vga11x19 font
	- some minor manpage and documentation fixes
	- fix Makefiles: data.c dependencies and src/tools/periph
	- verbose make install
	- debug output goes to ~/.dosemu/boot.log by default for dosemu.bin
	  too
	- allow $_dosmem in ~/.dosemurc and warn if more than 640K is used
	  for X or console graphics.
	- install midid by default
	- fix compiler warnings in msdos.h and comcom.c.
	From Stas
	- fix DOSEMU freeze (from ctrl+alt+pause) when $_rdtsc=(off)

+++ src/doc/tools/doSgmlTools.pl
@@ -285,7 +285,7 @@ sub convertToHTML {
 sub convertToText {
   my ($theFile) = @_;
   my ($theOutputFile) = &dirname($theFile) ."/" .&basename($theFile, ('.sgml', '.xml')) .".txt";
-  my ($theHTMLFile);
+  my ($theHTMLFile) = "";
 
   my ($command);
 
@@ -328,7 +328,9 @@ sub convertToText {
       print "Cleaning up temporary HTML file\n";
     }
     
-    unlink ($theHTMLFile);
+    if (length($theHTMLFile)) {
+      unlink ($theHTMLFile);
+    }
   }
 }
